+++ packages/expo-router/src/standard-navigation/index.tsx
@@ -210,7 +210,6 @@ function partitionNavigatorProps<
     screenLayout,
     screenListeners,
     screenOptions,
-    UNSTABLE_routeNamesChangeBehavior,
     UNSTABLE_router,
     ...extraProps
   } = props;
@@ -231,7 +230,6 @@ function partitionNavigatorProps<
     screenLayout,
     screenListeners,
     screenOptions,
-    UNSTABLE_routeNamesChangeBehavior,
     UNSTABLE_router,
   };
   return {